requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
SQLAlchemy==1.4.39
PyMySQL==1.0.2
python-dotenv
passlib==1.7.4
pydantic[email]
bcrypt==3.2.2